requirements.txt
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.2.0/en_core_web_md-3.2.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
Flask==2.0.3
Flask-RESTful==0.3.9
flatbuffers==2.0
gast==0.4.0
gunicorn==20.1.0
huggingface-hub==0.4.0
install==1.3.5
Keras-Preprocessing==1.1.2
langcodes==3.3.0
libclang==13.0.0
nltk==3.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pbr==5.8.1
pexpect==4.8.0
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.2.0/pt_core_news_sm-3.2.0-py3-none-any.whl
ptyprocess==0.7.0
PyDictionary==2.0.1
pyspellchecker==0.6.3
pyenchant==3.2.2
pure-eval==0.2.2
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spacy-wordnet==0.0.5
textblob==0.17.1
tqdm==4.62.3